Move the pcap job tests into a sibling file
jobs.rs reached 324 lines with the new test module inline, over the
repository's 300-line guard -- which is exactly what the guard is for. The
tests move to jobs/tests.rs, matching dispatch/tests.rs and
dispatch/policy_tests.rs beside them.
